# Conflicts:
#	powerfactory/powerfactory-converter/src/main/java/com/powsybl/powerfactory/converter/PowerFactoryImporter.java

+++ matpower/matpower-model/src/main/java/com/powsybl/matpower/model/MBranch.java
@@ -3,6 +3,7 @@
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ * SPDX-License-Identifier: MPL-2.0
  */
 package com.powsybl.matpower.model;
 
@@ -13,7 +14,7 @@ package com.powsybl.matpower.model;
  * @see <a href="https://matpower.org/doc/">https://matpower.org/doc/</a>
  * </p>
  *
- * @author Christian Biasuzzi <christian.biasuzzi@techrain.eu>
+ * @author Christian Biasuzzi {@literal <christian.biasuzzi@techrain.eu>}
  */
 public class MBranch {
 
@@ -70,7 +71,7 @@ public class MBranch {
     /**
      * initial branch status, 1 = in-service, 0 = out-of-service
      */
-    private double status;
+    private int status;
 
     /**
      * minimum angle difference, ThetaF - ThetaT (degrees)
@@ -162,11 +163,11 @@ public class MBranch {
         this.phaseShiftAngle = phaseShiftAngle;
     }
 
-    public double getStatus() {
+    public int getStatus() {
         return status;
     }
 
-    public void setStatus(double status) {
+    public void setStatus(int status) {
         this.status = status;
     }
